requirements.txt
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
colorama==0.4.4
decorator==4.4.2
Flask==2.0.1
gunicorn==20.1.0
idna==2.10
imageio==2.9.0
imageio-ffmpeg==0.4.4
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
moviepy==1.0.3
numpy==1.21.0
Pillow==8.2.0
proglog==0.1.9
qrcode==6.1
requests==2.25.1
six==1.16.0
tqdm==4.61.1
urllib3==1.26.6
Werkzeug==2.0.1
